+++ Models/KeychainService.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class KeychainService: NSObject {
+/*class KeychainService: NSObject {
 
   enum KeychainError: Error {
     case noPassword
@@ -64,3 +64,4 @@ class KeychainService: NSObject {
     return returned_string
   }
 }
+*/
